made cli commands blueprint
